feat: 문서화 및 테스트 작성

+++ src/main/java/com/example/board/service/PostService.java
@@ -4,6 +4,7 @@ import com.example.board.converter.PostConverter;
 import com.example.board.domain.Post;
 import com.example.board.domain.User;
 import com.example.board.dto.request.post.CreatePostRequest;
+import com.example.board.dto.request.post.DeletePostRequest;
 import com.example.board.dto.request.post.PostSearchCondition;
 import com.example.board.dto.request.post.UpdatePostRequest;
 import com.example.board.dto.response.PageResponse;
@@ -11,7 +12,6 @@ import com.example.board.dto.response.PostResponse;
 import com.example.board.exception.CustomException;
 import com.example.board.exception.ErrorCode;
 import com.example.board.repository.post.PostRepository;
-import com.example.board.repository.user.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -24,10 +24,10 @@ import org.springframework.transaction.annotation.Transactional;
 public class PostService {
 
     private final PostRepository postRepository;
-    private final UserRepository userRepository;
+    private final UserService userService;
 
     public Long createPost(CreatePostRequest requestDto) {
-        final User user = userRepository.findById(requestDto.authorId()).orElseThrow(() -> new CustomException(ErrorCode.USER_NOT_FOUND));
+        final User user = userService.getAvailableUser(requestDto.authorId());
         final Post post = postRepository.save(PostConverter.toPost(requestDto, user));
         return post.getId();
     }
@@ -52,8 +52,12 @@ public class PostService {
         post.update(requestDto.title(), requestDto.content());
     }
 
-    public void deletePost(Long id) {
-        final Post post = postRepository.findById(id).orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
+    public void deletePost(Long id, DeletePostRequest requestDto) {
+        final Post post = postRepository.findByIdWithAuthor(id).orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
+
+        if (!isAuthor(post, requestDto.authorId()))
+            throw new CustomException(ErrorCode.AUTHOR_NOT_MATCH);
+
         postRepository.delete(post);
     }
